+++ Rome.cabal
@@ -8,7 +8,7 @@ license-file:        LICENSE
 author:              Tommaso Piazza
 maintainer:          tommaso.piazza@gmail.com
 copyright:           2016 Tommaso Piazza
-category:            Web
+category:            Utility
 build-type:          Simple
 -- extra-source-files:
 cabal-version:       >=1.10
@@ -16,6 +16,10 @@ cabal-version:       >=1.10
 library
   hs-source-dirs:      src
   exposed-modules:     Lib
+                       , Data.Cartfile
+                       , Data.Romefile
+                       , Text.Parsec.Utils
+
   build-depends:       base >= 4.7 && < 5
                        , amazonka >= 1.4.1
                        , amazonka-s3 >= 1.4.1
